mul_neutral and pbs partly

+++ src/math/modular/module_switch.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 use proptest::prelude::*;
 
-pub fn rounded_div(dividend: u64, divisor: u64) -> u64 {
+pub fn rounded_div(dividend: u128, divisor: u128) -> u128 {
     // if dividend ^ divisor >= 0 {
     // println!("dividend: {:#x?}", dividend);
     // println!("divisor: {:#x?}", divisor);
@@ -12,8 +12,9 @@ pub fn rounded_div(dividend: u64, divisor: u64) -> u64 {
     // }
 }
 
-pub fn mod_switch(a: u64, old_q: u64, new_q: u64) -> u64 {
+pub fn mod_switch(a: u64, old_q: u128, new_q: u128) -> u64 {
     let nv: u128 = a as u128 * new_q as u128 / old_q as u128;
+    // let nv: u128 = a as u128 / rounded_div(old_q as u128, new_q as u128);
     nv as u64
 }
 
@@ -22,7 +23,7 @@ proptest! {
   #![proptest_config(ProptestConfig::with_cases(10000))]
   #[test]
   fn round_div_test(a in 1000..10000u64, b in any::<u64>().prop_filter("Not zero", |v| *v > 100000000000)){
-    rounded_div(a, b);
+    rounded_div(a as u128, b as u128);
   }
 }
 
@@ -36,9 +37,9 @@ proptest! {
 
 
 
-        let switched = mod_switch(source, higher_q, lower_q);
+        let switched = mod_switch(source, higher_q as u128, lower_q as u128);
         // prop_assert!(switched > 0)
-        let unswitched = mod_switch(switched, lower_q, higher_q);
+        let unswitched = mod_switch(switched, lower_q as u128, higher_q as u128);
        // prop_assert_eq!(unswitched, source);
        let delta = 5;
        prop_assert!(unswitched >= if source.wrapping_sub(delta)>source {0} else {source.wrapping_sub(delta)}  && unswitched < source.wrapping_add(delta), "unswitched: {unswitched}, source: {source} ");
